this adds support for string featuers in python as follows:
>>> from shogun.Features import *
>>> x=StringCharFeatures(Alphabet(DNA))
>>> x.set_string_features(['blafoo','hallo','dummdidumm'])
>>> x.select_feature_vector(1)
>>> x.get_str()
array(['h', 'a', 'l', 'l', 'o'], 
			  dtype='|S1')

there is a certain overhaul in the WD* kernels necessary -
as the shift kernel is not yet really supported (and we are losing
		control if we go on duplicating the code like this...)

+++ src/kernel/WeightedDegreePositionStringKernel.cpp
@@ -41,7 +41,7 @@ struct S_THREAD_PARAM
 	DREAL* sqrtdiag_rhs;
 };
 
-CWeightedDegreePositionStringKernel::CWeightedDegreePositionStringKernel(LONG size, DREAL* w, INT d, 
+CWeightedDegreePositionStringKernel::CWeightedDegreePositionStringKernel(INT size, DREAL* w, INT d, 
 		INT max_mismatch_, INT * shift_, 
 		INT shift_len_, bool use_norm,
 		INT mkl_stepsize_)
